fix: INT4 GEMM zero-point offset (+1) and FP16 scales
Two fixes for INT4 quantized weight dequantization:

1. Zero-point offset: AutoRound stores zero points with a -1 bias.
   The actual zero = stored_value + 1 (8 for symmetric INT4).
   This was the primary accuracy bug causing ~1% per-layer cos error
   that compounded to 68° angular drift over 64 layers.

2. FP16 scales: Preserve scales as FP16 (10 mantissa bits) instead
   of truncating to BF16 (7 mantissa bits) during upload.

Results:
- Layer 0 output cos: 0.991 → 0.9996
- Layer 63 output cos: 0.317 → 0.917 (was garbage, now near-correct)
- Prefill logits: token 760 moved from rank 11996 to rank 1
- Expected token 248068 moved from rank 12 to rank 8

+++ crates/backends/native/src/engine.rs
@@ -23,7 +23,7 @@ use crate::gdn::GdnState;
 
 use infers_kv::PagedKvManager;
 
-use half::bf16;
+use half::{bf16, f16};
 use infers_kv::KvCacheDtype;
 use infers_cuda::CudaSlice;
 
@@ -410,7 +410,7 @@ impl ForwardEngine {
         config: &Int4GemmConfig,
         output: &mut CudaSlice<half::bf16>,
         weight: &CudaSlice<u32>,
-        scales: &CudaSlice<half::bf16>,
+        scales: &CudaSlice<f16>,
         zeros: &CudaSlice<u32>,
         input: &CudaSlice<half::bf16>,
     ) -> Result<()> {
